+++ src/test/java/kitchenpos/menu/tobe/domain/MenuServiceTest.java
@@ -54,4 +54,18 @@ class MenuServiceTest {
 
         assertFalse(menu.isMenuDisplayStatus());
     }
+
+    @Test
+    @DisplayName("메뉴상품 금액 총 합보다 메뉴 가격이 비싸지면 메뉴는 숨겨진다.")
+    void priceFail4() {
+        final var menu = new Menu(menuName.create("메뉴이름"), new MenuPrice(10_000L), menuGroup, true, menuProducts);
+
+        assertTrue(menu.isMenuDisplayStatus());
+
+        productRepository.save(product);
+        menuRepository.save(menu);
+        menuService.syncMenuDisplayStatisWithProductPrices(product.getId(), BigDecimal.valueOf(5_000L));
+
+        assertFalse(menu.isMenuDisplayStatus());
+    }
 }
